requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.25
apscheduler==3.10.4
python-dotenv==1.0.0